fix(money): update CustomOrderResponse getter return type
Compile broke on CI: getter still declared Float while the field moved
to BigDecimal. My earlier replace_all only matched "Float priceSum"
patterns, so the getter signature was missed.

+++ src/main/java/com/albertkingdom/shoppingwebsite/model/CustomOrderResponse.java
@@ -29,7 +29,7 @@ public class CustomOrderResponse {
         this.id = id;
     }
 
-    public Float getPriceSum() {
+    public BigDecimal getPriceSum() {
         return priceSum;
     }
